requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
sqlmodel==0.0.24
python-dotenv==1.0.1
psycopg2-binary==2.9.10